requirements.txt
playwright==1.49.1
google-generativeai==0.8.4
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
